requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
apify-client==1.7.0
pandas==2.1.3
gender-guesser==0.4.0
supabase==2.9.0
pyairtable==2.3.3

# Production server
gunicorn==21.2.0

# === Async Task Queue ===
celery==5.3.4
redis==4.6.0
celery[redis]==5.3.4
kombu==5.3.4

# Rate limiting and security
flask-limiter==3.5.0

# Error tracking
sentry-sdk[flask]==1.40.0

# Data validation
marshmallow==3.20.1
